refactor(core): fold single-consumer request modules into model-request
